Added the aperiodical hat
See https://arxiv.org/abs/2303.10798

This would have gone a whole lot easier if I'd just read the
documentation of the package and used `\DefineTile` from the outset!

The hat is used in both right-side up and flipped, so for now this is
implemented as two tiles.  To make it slightly simpler, `\DefineTile`
now takes a star option to indicate that the tile should be considered
as defined clockwise instead of anticlockwise.  This affects placement
of tiles along edges.

Also, the longer edge should be considered as two half-edges.