Better transformation code
The code that transforms a tile to align with another now also applies a
scaling so that the matched sides are the same length.  The means that a
scale factor applied to the first tile propogates through a pattern,
which is much better than trying to work out how the individual tiles
should be scaled compatibly with their alignment transformations.

Also, the super tiles are now implemented for the aperiodical hat
tiling.