LMS system for the clusters implemented
This works on a neighbour-basis rather than a positional basis, so needs
enough rules to cope with not just tile replacement but
tile-with-neighbour replacement.  Fortunately, there's a fair amount
that can be leveraged with templates, and even more fortunately my first
guesses tended to be right so I didn't have to do too much thinking!